+++ R/IRanges.R
@@ -1,0 +1,11 @@
+#' Integer ranges
+#'
+#' Examples `IRanges` object.
+#'
+#' @examples
+#' if (requireNamespace("IRanges", quietly = TRUE)) {
+#'     data(IRanges)
+#'     class(IRanges)
+#'     IRanges
+#' }
+"IRanges"