+++ R/data.frame.R
@@ -1,0 +1,9 @@
+#' Data frame
+#'
+#' Example `data.frame` object.
+#'
+#' @examples
+#' data(data.frame)
+#' class(data.frame)
+#' data.frame
+"data.frame"