+++ R/matrix_lfc.R
@@ -1,0 +1,9 @@
+#' Log fold change
+#'
+#' Example `matrix` containing log base 2 fold change values.
+#'
+#' @examples
+#' data(matrix_lfc)
+#' class(matrix_lfc)
+#' matrix_lfc
+"matrix_lfc"