+++ R/sparseMatrix.R
@@ -1,0 +1,11 @@
+#' Sparse matrix
+#'
+#' Example `sparseMatrix` object.
+#'
+#' @examples
+#' if (requireNamespace("Matrix", quietly = TRUE)) {
+#'     data(sparseMatrix)
+#'     class(sparseMatrix)
+#'     sparseMatrix
+#' }
+"sparseMatrix"