+++ R/tbl_df.R
@@ -1,0 +1,11 @@
+#' Tibble
+#'
+#' Example `tbl_df` object.
+#'
+#' @examples
+#' if (requireNamespace("tibble", quietly = TRUE)) {
+#'     data(tbl_df)
+#'     class(tbl_df)
+#'     tbl_df
+#' }
+"tbl_df"