+++ data-raw/data.table.R
@@ -1,0 +1,5 @@
+library(usethis)
+library(data.table)
+data(matrix)
+data.table <- as.data.table(matrix, keep.rownames = TRUE)
+use_data(data.table, compress = "xz", overwrite = TRUE)