+++ data-raw/tbl_df.R
@@ -1,0 +1,5 @@
+library(usethis)
+library(tibble)
+data(matrix)
+tbl_df <- as_tibble(matrix, rownames = "rowname")
+use_data(tbl_df, compress = "xz", overwrite = TRUE)